R6 class that creates b-splines.

Creates b-splines given some set of knot points, and then evaluates the splines given some set of coefficients for the splines. It has 3 arguments initialisation : f = BSpline(K, t_knot, m).

Public fields

t_knot

Spline knot points.

K

Order of polynomial.

m

Spline coefficients (\(m = M \cdot D\)).

x_mean

If set, these will be used to scale the output (default to 0).

x_std

\(x_{out} = x_{std} \cdot (X \cdot m) + x_{mean}\)

t_pp

pp break points (size(t_pp) = length(t_knot) - 2*K + 1).

C

Piecewise polynomial coefficients (size(C) = [length(t_pp)-1, K]).

Methods

Public methods

bspline$new()

bspline$clone()

Method new()

Instantiate an object of the class bspline.

Usage

bspline$new(t_knot, K, m)

Arguments

t_knot

Spline knot points.

K

Order of polynomial.

m

Spline coefficients (\(m = M \cdot D\)).

Method clone()

The objects of this class are cloneable with this method.

Usage

bspline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.